Clear content related maps before the processing of new test class
